move SessionProvider into providers folder
