+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import './globals.css';
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
-import { SessionProvider } from '@/app/(client)/(auth)/components/SessionProvider';
+import { SessionProvider } from './(providers)/SessionProvider';
 import { getServerSession } from 'next-auth';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import Providers from './(providers)/QueryClientProvider';
@@ -9,8 +9,8 @@ import Providers from './(providers)/QueryClientProvider';
 const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'spotifygpt',
+  description: 'use AI to discover music',
 };
 
 export default async function RootLayout({
